requirements.txt
beir>=1.0.0
rank-bm25
sentence-transformers
faiss-cpu
torch
tqdm